requirements.txt
bs4
gig-nuuuwan
matplotlib
numpy 
Pillow
utils-nuuuwan
wordcloud